ecommerce_blueprint_version/requirements.txt
bcrypt==3.1.7
cement==2.8.2
certifi==2020.12.5
cffi==1.14.0
chardet==4.0.0
Click==7.0
colorama==0.4.3
distlib==0.3.1
filelock==3.0.12
flask==1.1.2
Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
# future==0.16.0
gunicorn==19.9.0
html5lib==1.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.1
jmespath==0.10.0
MarkupSafe==1.1.1
mechanize==0.4.5
Pillow==8.2.0
pycparser==2.19
PyJWT==1.7.1
python-dateutil==2.8.1
requests==2.25.1
six==1.14.0
soupsieve==2.1
SQLAlchemy==1.3.22
urllib3==1.25.11
webencodings==0.5.1
Werkzeug==1.0.0
WTForms==2.3.3
Flask-Bcrypt==0.7.1
twilio==6.55.0